+++ changelog.md
@@ -8,6 +8,7 @@
 * Fix crash when creating a music without playing it immediately (#1896).
 * Fix crash when calling `sol.main.get_game()` while restarting (#1904).
 * Fix possible duplicate name when creating an entity dynamically (#1908).
+* Fix OpenAL error A003 when stopping a music that was just started.
 * Fix index offsets for channels of IT music files.
 * Fix possible error when retrieving joypad name (#1895).
 * Add hero argument to `item:on_using()` event (#1899).